Fix some cases for addition with numbers of differing signs
There's not much difference here; we apply two's complement conversion to rhs if the signs differ and add a check for the mantissa being zero after the sum is calculated in addition to the one we already had for the exponents, and everything just works more or less, which is rad. Still loads of cases not yet covered, so we'll see if this strategy is something we'll continue with in the end or not.